protocol: create new protocol module
Create a new protocol module from the server module (basically
just renaming it) and move a bunch of functions that don't really
belong into the vfs module as well.
