+++ test/int/test_protocol.py
@@ -1,0 +1,17 @@
+
+from __future__ import absolute_import, print_function
+from io import BytesIO
+import sys
+
+from wvpytest import *
+
+from bup import git, metadata, vfs, protocol
+
+
+def test_item_read_write():
+    x = vfs.Root(meta=13)
+    stream = BytesIO()
+    protocol.write_item(stream, x)
+    print('stream:', repr(stream.getvalue()), stream.tell(), file=sys.stderr)
+    stream.seek(0)
+    wvpasseq(x, protocol.read_item(stream))
